fetch/metadata/generated/element-a.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/element-a.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for HTML "a" element navigation</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, {test, userActivated, attributes}) {
    const win = window.open();
    const anchor = win.document.createElement('a');
    anchor.setAttribute('href', url);

    for (const [ name, value ] of Object.entries(attributes)) {
      anchor.setAttribute(name, value);
    }

    win.document.body.appendChild(anchor);

    test.add_cleanup(() => win.close());

    if (userActivated) {
      test_driver.bless('enable user activation', () => anchor.click());
    } else {
      anchor.click();
    }
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      makeRequestURL(key, ['httpOrigin'], {mime: 'text/html'}),
      {
        test: t,
        userActivated: false,
        attributes: {}
      }
    );

    // `induceRequest` does not necessarily trigger a navigation, so the Python
    // handler must be polled until it has received the initial request.
    return retrieve(key, {poll: true})
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy same-origin destination - no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      makeRequestURL(key, ['httpSameSite'], {mime: 'text/html'}),
      {
        test: t,
        userActivated: false,
        attributes: {}
      }
    );

    // `induceRequest` does not necessarily trigger a navigation, so the Python
    // handler must be polled until it has received the initial request.
    return retrieve(key, {poll: true})
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy same-site destination - no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      makeRequestURL(key, ['httpCrossSite'], {mime: 'text/html'}),
      {
        test: t,
        userActivated: false,
        attributes: {}
      }
    );

    // `induceRequest` does not necessarily trigger a navigation, so the Python
    // handler must be polled until it has received the initial request.
    return retrieve(key, {poll: true})
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy cross-site destination - no attributes');
  </script>
  </body>
</html>
